[MOV] im_livechat: move CCThrottleFunction outside of public_livechat.js
Task-2212347
